backend/requirements.txt
flask==3.1.0
flask-cors==6.0.2
anthropic==0.75.0
requests==2.32.5
yt-dlp==2025.12.8
certifi==2025.11.12
gunicorn==23.0.0
python-dotenv==1.1.0
